share in-memory exporter between test cases
instead of creating a new exporter in RumRailsTest and
RackRailsTest have EXPORTER instance created globally
and reset it in the teardown method of the testcases.

It may be useful to figure out in the future why creating
a new exporter in each testcase was resulting in no spans
making it to the exporter in the second test to run.